# docs/swagger.yaml
definitions:
  github_com_evrone_go-clean-template_internal_controller_http_v1_request.AdminSubjectCreateRequest:
    properties:
      exam:
        $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.ExamCategory'
      name:
        type: string
    required:
    - exam
    - name
    type: object
  github_com_evrone_go-clean-template_internal_controller_http_v1_request.AdminTopicCreateRequest:
    properties:
      name:
        type: string
      subjectId:
        type: string
    required:
    - name
    - subjectId
    type: object
...
  github_com_evrone_go-clean-template_internal_entity.AdminReferralSummary:
    properties:
      newUsers:
        type: integer
      range:
        type: string
      rewardsPaid:
        type: integer
      totalReferrals:
        type: integer
    type: object
...
  github_com_evrone_go-clean-template_internal_entity.ExamCategory:
    enum:
    - NEET_PG
    - NEET_UG
    - JEE
    - UPSC
    type: string
    x-enum-varnames:
    - ExamCategoryNEETPG
    - ExamCategoryNEETUG
    - ExamCategoryJEE
    - ExamCategoryUPSC
...
  github_com_evrone_go-clean-template_internal_entity.Question:
    properties:
      choiceType:
        $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.QuestionChoiceType'
      correctOption:
        type: integer
      difficultyLevel:
        type: integer
      exam:
        $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.ExamCategory'
      explanation:
        type: string
      id:
        type: string
      isActive:
        type: boolean
      isClinical:
        type: boolean
      isHighYield:
        type: boolean
      isImageBased:
        type: boolean
      optionA:
        type: string
      optionB:
        type: string
      optionC:
        type: string
      optionD:
        type: string
      questionText:
        type: string
      subjectId:
        type: string
      topicId:
        type: string
    type: object
  github_com_evrone_go-clean-template_internal_entity.QuestionChoiceType:
    enum:
    - single
    - multi
    type: string
    x-enum-varnames:
    - ChoiceTypeSingle
    - ChoiceTypeMulti
  github_com_evrone_go-clean-template_internal_entity.QuestionCreateRequest:
    properties:
      choiceType:
        $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.QuestionChoiceType'
      correctOption:
        maximum: 4
        minimum: 1
        type: integer
      difficultyLevel:
        type: integer
      exam:
        $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.ExamCategory'
      explanation:
        type: string
      isActive:
        type: boolean
      isClinical:
        type: boolean
      isHighYield:
        type: boolean
      isImageBased:
        type: boolean
      optionA:
        type: string
      optionB:
        type: string
      optionC:
        type: string
      optionD:
        type: string
      questionText:
        type: string
      subjectId:
        type: string
      topicId:
        type: string
    required:
    - correctOption
    - exam
    - optionA
    - optionB
    - optionC
    - optionD
    - questionText
    - subjectId
    - topicId
    type: object
  github_com_evrone_go-clean-template_internal_entity.QuestionUpdateRequest:
    properties:
      choiceType:
        $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.QuestionChoiceType'
      correctOption:
        type: integer
      difficultyLevel:
        type: integer
      explanation:
        type: string
      isActive:
        type: boolean
      isClinical:
        type: boolean
      isHighYield:
        type: boolean
      isImageBased:
        type: boolean
      optionA:
        type: string
      optionB:
        type: string
      optionC:
        type: string
      optionD:
        type: string
      questionText:
        type: string
      subjectId:
        type: string
      topicId:
        type: string
    type: object
...
  github_com_evrone_go-clean-template_internal_entity.Subject:
    properties:
      exam:
        $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.ExamCategory'
      id:
        type: string
      isActive:
        type: boolean
      name:
        type: string
    type: object
...
  github_com_evrone_go-clean-template_internal_entity.Topic:
    properties:
      id:
        type: string
      isActive:
        type: boolean
      name:
        type: string
      subjectId:
        type: string
    type: object
...
  internal_controller_http_v1.ErrorResponse:
    properties:
      error:
        example: message
        type: string
    type: object
info:
  contact: {}
paths:
...
  /admin/questions:
    get:
      parameters:
      - description: Exam
        in: query
        name: exam
        type: string
      - description: Subject ID
        in: query
        name: subjectId
        type: string
      - description: Topic ID
        in: query
        name: topicId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.Question'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      security:
      - AdminAuth: []
      summary: List questions
      tags:
      - 'Admin: Questions'
    post:
      consumes:
      - application/json
      parameters:
      - description: Question payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.QuestionCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.Question'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      security:
      - AdminAuth: []
      summary: Create question
      tags:
      - 'Admin: Questions'
  /admin/questions/{id}:
    delete:
      parameters:
      - description: Question ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      security:
      - AdminAuth: []
      summary: Delete question
      tags:
      - 'Admin: Questions'
    get:
      parameters:
      - description: Question ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.Question'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      security:
      - AdminAuth: []
      summary: Get question
      tags:
      - 'Admin: Questions'
    patch:
      consumes:
      - application/json
      parameters:
      - description: Question ID
        in: path
        name: id
        required: true
        type: string
      - description: Update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.QuestionUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.Question'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      security:
      - AdminAuth: []
      summary: Update question
      tags:
      - 'Admin: Questions'
  /admin/referrals/summary:
    get:
      parameters:
      - description: today|7d|30d
        in: query
        name: range
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.AdminReferralSummary'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      security:
      - AdminAuth: []
      summary: Referral summary
      tags:
      - 'Admin: Referrals'
  /admin/subjects:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.Subject'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      security:
      - AdminAuth: []
      summary: List subjects
      tags:
      - 'Admin: Subjects'
    post:
      consumes:
      - application/json
      parameters:
      - description: Subject payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_evrone_go-clean-template_internal_controller_http_v1_request.AdminSubjectCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.Subject'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      security:
      - AdminAuth: []
      summary: Create subject
      tags:
      - 'Admin: Subjects'
  /admin/topics:
    get:
      parameters:
      - description: Subject ID
        in: query
        name: subjectId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.Topic'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      security:
      - AdminAuth: []
      summary: List topics
      tags:
      - 'Admin: Topics'
    post:
      consumes:
      - application/json
      parameters:
      - description: Topic payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_evrone_go-clean-template_internal_controller_http_v1_request.AdminTopicCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.Topic'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      security:
      - AdminAuth: []
      summary: Create topic
      tags:
      - 'Admin: Topics'
...
  /subjects:
    get:
      parameters:
      - description: Exam category
        in: query
        name: exam
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.Subject'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      security:
      - UserAuth: []
      summary: List subjects for current exam
      tags:
      - 'App: User'
  /topics:
    get:
      parameters:
      - description: Subject ID
        in: query
        name: subjectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.Topic'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      security:
      - UserAuth: []
      summary: List topics for subject
      tags:
      - 'App: User'
...
swagger: "2.0"